Added simple dashboard template + restructured api
