+++ assets/api/views/Client.py
@@ -1,25 +1,17 @@
 from ast import parse
 from flask_apispec import marshal_with, doc, use_kwargs
 from flask_apispec.views import MethodResource
-from marshmallow import Schema, fields
 
 from flask_restful import Resource, reqparse
 
 from assets.classes.ClientClass import Client
+from ..models import ClientGETRequestModel, ClientPOSTRequestModel, InvalidResponseModel
+
 from main import clients_queue
 
 
 parser = reqparse.RequestParser()
 parser.add_argument('id')
-
-class ClientGETRequestModel(Schema):
-    id = fields.Str()
-    order_number = fields.Int()
-    timestamp = fields.Str()
-
-
-class ClientPOSTRequestModel(Schema):
-    id = fields.Str()
 
 
 class ClientEndpoint(MethodResource, Resource):
@@ -48,7 +40,10 @@ class ClientEndpoint(MethodResource, Resource):
         tags=['Client']
     )
     @use_kwargs(ClientPOSTRequestModel)
-    @marshal_with(ClientGETRequestModel)
+    @marshal_with(ClientGETRequestModel, code=201)
+    @marshal_with(InvalidResponseModel, code=400)
+    @marshal_with(InvalidResponseModel, code=404)
+    @marshal_with(InvalidResponseModel, code=500)
     def post(self, **kwargs):
         """
         POST Method - Endpoint used to get info about specific client.
