+++ assets/api/views/Status.py
@@ -1,18 +1,12 @@
 from flask_apispec import marshal_with, doc
 from flask_apispec.views import MethodResource
-from marshmallow import Schema, fields
 
 from flask_restful import Resource
 
-from .Client import ClientGETRequestModel
+from ..models import QueueModel, InvalidResponseModel
 
 from main import clients_queue
 
-class QueueModel(Schema):
-    queue_size = fields.Int()
-    clients = fields.Nested(
-        ClientGETRequestModel(many=True)
-    )
 
 class StatusEndpoint(MethodResource, Resource):
 
@@ -20,7 +14,10 @@ class StatusEndpoint(MethodResource, Resource):
         description='This endopoint is used to fetch current queue',
         tags=['Queue Status']
     )
-    @marshal_with(QueueModel)
+    @marshal_with(QueueModel, code=201)
+    @marshal_with(InvalidResponseModel, code=400)
+    @marshal_with(InvalidResponseModel, code=404)
+    @marshal_with(InvalidResponseModel, code=500)
     def get(self):
         clients = clients_queue.get_iterable()
